Update makeHessPositiveDefinite
This new method is a lot better, but it can fail due to rounded eigenvectors not having the correction effect that they should. With the bending energy Hessians, complex eigenvalues are sometimes found it should be investigated whether that's an error in the algorithm or in the bending energy Hessian itself.